Fix #18 problem should be fixed now. Update of pygccxml and pyplusplus is necessary .. best way is to delete it in thirdparty

+++ python/hand_made_wrappers.py
@@ -136,9 +136,9 @@ WRAPPER_DEFINITION_R3Vector =\
 
 PyObject * R3Vector_getArray(GIMLI::R3Vector & vec){
     import_array2("Cannot import numpy c-api from pygimli hand_make_wrapper2", NULL);
-    npy_intp length = (size_t)vec.size();
+    npy_intp length = (ssize_t)vec.size();
       
-    ssize_t dim2 [] = {length, 3};
+    int dim2 [] = {length, 3};
     PyObject * ret = PyArray_SimpleNew(2, dim2, NPY_DOUBLE);
 
     // check if array is contiguous here
